Add files via upload
BGC-Backtest_BASE.ipynb is a Jupyter Notebook file written in Python3 that creates a backtesting program dynamically in Python.  This is the base code that will be expanded out by further building out  a trading dictionary and strategy dictionary to allow for easy copy/paste code for TA and backtests as well as portfolio optimization.